bug: State stores loose state when tasks are reassigned under EOS with standby replicas and default acceptable lag. The stream attached in the test produces a unique id per observed value and passes the value and id to the outgoing topic. The stream rely on:  * Two state stores.  * EXACTLY_ONCE_V2 semantics  * One Standby replica.
The test however demonstrates that a re-balancing event can break the exactly-once semantics. The issue can be reliably reproduced on a on an i7-8750H CPU @ 2.20GHz × 12 with 32 GiB Memory when using caching and the default acceptable lag for standby replicas.

When caching is disabled OR the acceptable lag is set to 0 then the test no longer breaks.
The underlying cause is unknown and so is whether or not the above settings fixes or hides the problem.

A similar issue have been reported here:
 https://stackoverflow.com/questions/69038181/kafka-streams-aggregation-data-loss-between-instance-restarts-and-rebalances

Flow :
Produce 3000 messages.
Start a stream.
Wait for the 3000 messages to be processed.
Start a new stream and wait for it to start syncing
Produce 60.000 messages
Wait for 5 second.
Start a new thread which should introduce a rebalancing event.
Wait until the entire log is processed by the stream.
Check the uniqueness of the assigned Ids.